Refactor(package): move AuthModuleService to impl
